Move tempo_deduce content into the already existing utils file
